+++ packages/transaction-controller/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Fix `isFirstTimeInteraction` to be correctly settled when transferring ERCX tokens. ([#6686](https://github.com/MetaMask/core/pull/6686))
+- Update `isFirstTimeInteraction` to be determined using recipient if token transfer. ([#6686](https://github.com/MetaMask/core/pull/6686))
 
 ## [60.4.0]